calc subtotal na pagina estava trabalhando com int, mudei para float

+++ src/main/webapp/pedidos/pedidos2.js
@@ -442,7 +442,7 @@ function simularTotal() {
   var total = 0;
   $("#pedidoItensBody").children('tr').each(function(i){
     var subtotal = $(this).find(".itemSubtotal").val();
-    total += parseInt(subtotal);
+    total += parseFloat(subtotal);
   });
   $("#total").val(total);
 }
